Add support for rename
Implement rename by first duplicating the metadata for the destination,
then executing the rename and finally deleting the metadata for
the source location. This prevents incomplete listings for both,
the source and the destination as clients will fail or retry on
missing objects. In case of invalid operations, the rename
will return false and the duplicated metadata will be deleted again.
In case of errors, rename will throw Exceptions and the metadata
will not be cleaned up, thus preventing incomplete listings for both,
source and destination. Manual cleanup will be required in these
scenarios but preventing reading incomplete data is the most important
right now.